+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Added repeat Every 2 Weeks support (please update your events, the IDs for the others have been changed.  Sorry).
 - Added "Enable drag & drop for events" option to the Configuration dialog.
 - Renamed the tabs and setting names for the Configuration dialog (smaller and better).
+- Added "Organizer" to the Configuration dialog, with support to change the name and email address.
 - 
 
 ### Version 0.9.6: